Move `brokerd` service task to root `.data` mod
